pyproject.toml
[project]
name = "emoji-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.4",
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.0.1",
]
